Reorganize frontend modules and other files
strcalc/src/main/frontend was becoming a bit cluttered. This should make
things a bit more clear and manageable.
