+++ strcalc/src/main/frontend/ci/vitest.config.js
@@ -1,5 +1,5 @@
 import { defineConfig, mergeConfig } from 'vitest/config'
-import viteConfig from './vite.config'
+import viteConfig from '../vite.config'
 
 export default mergeConfig(viteConfig, defineConfig({
   test: {
