+++ strcalc/src/main/frontend/plugins/rollup-plugin-handlebars-precompiler.js
