+++ strcalc/src/main/frontend/test/helpers.js
